pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
stripe = "^10.12.0"
forex-python = "^1.8"
django-extensions = "^3.2.3"
flake8 = "^7.1.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
